+++ demo/lib/demo_web/router.ex
@@ -24,7 +24,9 @@ defmodule DemoWeb.Router do
 
     live_session :backpex_admin,
       on_mount: [{DemoWeb.LiveUserAuth, :live_user_demo}, Backpex.InitAssigns] do
+      live_resources("/authors", AuthorLive)
       live_resources("/posts", PostLive)
+      live_resources("/comments", CommentLive)
     end
   end
 end